Task-2/api/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
prometheus_client